requirements.txt
altgraph==0.17.2
asgiref==3.7.2
asttokens==2.0.8
auto-py-to-exe==2.22.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
bottle==0.12.23
bottle-websocket==0.2.9
captcha==0.5.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.5
crispy-bootstrap5==0.7
cryptography==40.0.2
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==4.0.0
Django==4.2.3
django-allauth==0.52.0
django-cors-headers==4.2.0
django-countries==7.5.1
django-crispy-forms==1.14.0
django-cuser==2017.3.16
django-filebrowser==4.0.3
django-grappelli==3.0.6
django-hosts==5.2
django-model-utils==4.3.1
django-modeltranslation==0.19.7
django-recaptcha==3.0.0
django-rest-auth==0.9.5
django-social-share==2.3.0
django-templated-mail==1.1.1
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.4.1
drf-yasg==1.21.7
Eel==0.14.0
entrypoints==0.4
executing==0.10.0
exif==1.6.0
filelock==3.8.0
flake8==6.0.0
Flask==2.3.2
future==0.18.3
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.4
imageio==2.21.2
imageio-ffmpeg==0.4.7
inflection==0.5.1
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
lxml==4.9.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
moviepy==1.0.3
mysqlclient==2.2.4
nest-asyncio==1.5.5
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pandas==2.2.3
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2022.9.24
platformdirs==2.5.2
plum-py==0.8.6
proglog==0.1.10
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pygame==2.1.2
Pygments==2.13.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-whois==0.8.0
python3-openid==3.2.0
pytube==12.1.0
pytz==2023.3
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0.1
pyzmq==23.2.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.4.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.3.6
whichcraft==0.6.1
whitenoise==6.5.0
windows-curses==2.3.0
yattag==1.15.1
zope.event==4.5.0
zope.interface==5.4.0
